requirements-dev.txt
pytest
httpx
ruff
black==23.11.0
pandas-stubs
pytest-playwright
